requirements.txt
astroid==3.3.8
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dill==0.3.9
fonttools==4.56.0
idna==3.10
iqoptionapi @ file:///media/cleytonfs/21db2a91-5ee2-437c-a5c2-f86c05d9cc073/home/cleyton/Documentos/PROJETOS2025/BERICH_BOT/berich2.0/iqoptionapi
isort==6.0.1
kiwisolver==1.4.8
matplotlib==3.10.1
mccabe==0.7.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pyaes==1.6.1
pylint==3.3.4
pyparsing==3.2.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
TgCrypto==1.2.5
tomlkit==0.13.2
tzdata==2024.2
urllib3==2.3.0
websocket-client==0.56.0
